% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.r
\name{amv_dim}
\alias{amv_dim}
\title{Dimensions.}
\usage{
amv_dim(x)
}
\arguments{
\item{x}{array, matrix or vector}
}
\description{
Consistent dimensions for vectors, matrices and arrays.
}
\keyword{internal}
