% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.r
\name{amv_dimnames}
\alias{amv_dimnames}
\title{Dimension names.}
\usage{
amv_dimnames(x)
}
\arguments{
\item{x}{array, matrix or vector}
}
\description{
Consistent dimnames for vectors, matrices and arrays.
}
\details{
Unlike \code{\link{dimnames}} no part of the output will ever be
null.  If a component of dimnames is omitted, \code{amv_dimnames}
will return an integer sequence of the appropriate length.
}
\keyword{internal}
