\name{as.list.split}
\alias{as.list.split}
\title{Convert split list to regular list}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Strip off label related attributed to make a strip list as regular list
}
\usage{as.list.split(x, ...)}
\arguments{
\item{x}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
