% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.list.split}
\alias{as.list.split}
\title{Convert split list to regular list.}
\usage{
\method{as.list}{split}(x, ...)
}
\arguments{
  \item{x}{object to convert to a list}

  \item{...}{unused}
}
\description{
Strip off label related attributed to make a strip list as regular list
}
\keyword{internal}

