\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create progress bar}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create progress bar object from text string.
}
\usage{create_progress_bar(name = "none")}
\arguments{
\item{name}{type of progress bar to create}
\item{}{progress}
}

\details{}

\examples{}

