\name{defaults}
\alias{defaults}
\title{Set defaults}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convient method for combining a list of values with their defaults.
}
\usage{defaults(x, y)}
\arguments{
\item{x}{list of values}
\item{y}{defaults}
}

\details{}

\examples{}
\keyword{manip}
