% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.r
\name{defaults}
\alias{defaults}
\title{Set defaults.}
\usage{
defaults(x, y)
}
\arguments{
\item{x}{list of values}

\item{y}{defaults}
}
\description{
Convient method for combining a list of values with their defaults.
}
\keyword{manip}
