\name{desc}
\alias{desc}
\title{Descending order...}
\usage{desc(x)}

\description{
  Descending order
}

\details{
  Transform a vector into a format that will be sorted in descending
  order.
}
\keyword{internal}
\keyword{desc(1:10)}
\keyword{desc(factor(letters))}
\keyword{first_day}
\keyword{<-}
\keyword{seq(as.Date("1910/1/1"),}
\keyword{as.Date("1920/1/1"),}
\keyword{"years")}
\keyword{desc(first_day)}
\arguments{
  \item{x}{vector to transform}
}
