% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{empty}
\alias{empty}
\title{Check if a data frame is empty.}
\usage{
empty(df)
}
\arguments{
\item{df}{data frame to check}
}
\description{
Empty if it's null or it has 0 rows or columns
}
\keyword{internal}
