\name{eval.quoted}
\alias{eval.quoted}
\title{Evaluate a quoted list of variables}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Evaluates quoted variables in specified environment
}
\usage{eval.quoted(exprs,  envir = parent.frame(), enclos = if (is.list(envir) || is.pairlist(envir)) parent.frame() else baseenv())}
\arguments{
\item{exprs}{}
\item{envir}{}
\item{enclos}{}
}
\value{a list}
\details{}

\examples{}
\keyword{internal}
