\name{failwith}
\alias{failwith}
\title{Fail with}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Modify a function so that it returns a default value when there is an error.
}
\usage{failwith(default = NULL, f, quiet = FALSE)}
\arguments{
\item{default}{default value}
\item{f}{function}
\item{quiet}{}
}
\value{a function}
\details{}
\seealso{\code{\link{try_default}}}
\examples{}

