% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{[.split}
\alias{[.split}
\title{Subset splits.}
\usage{
\method{[}{split}(x, i, ...)
}
\arguments{
\item{x}{split object}

\item{i}{index}

\item{...}{unused}
}
\description{
Subset splits, ensuring that labels keep matching
}
\keyword{internal}
