\name{indexed_array}
\alias{indexed_array}
\alias{[[.indexed_array}
\alias{names.indexed_array}
\alias{length.indexed_array}
\title{An indexed array}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a indexed array, a space efficient way of indexing into a large array
}
\usage{indexed_array(env, index)}
\arguments{
\item{env}{environment containing data frame}
\item{index}{}
}

\details{}

\examples{}
\keyword{internal}
