\name{indexed_df}
\alias{indexed_df}
\alias{length.indexed}
\alias{names.indexed}
\alias{as.list.indexed}
\alias{[[.indexed_df}
\alias{[.indexed}
\alias{print.indexed}
\title{An indexed list}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a indexed list, a space efficient way of indexing into a large data frame
}
\usage{indexed_df(env, index)}
\arguments{
\item{env}{environment containing data frame}
\item{index}{}
}

\details{}

\examples{}
\keyword{internal}
