\name{indexed_list}
\alias{indexed_list}
\alias{length.indexed_list}
\alias{names.indexed_list}
\alias{as.list.indexed_list}
\alias{[[.indexed_list}
\title{An indexed list}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a indexed list, a space efficient way of indexing into a large data frame
}
\usage{indexed_list(env, index)}
\arguments{
\item{env}{environment containing data frame}
\item{index}{}
}

\details{}

\examples{}

