\name{indexed_array}
\title{An indexed array.}
\usage{indexed_array(env, index)}

\description{
  An indexed array. Create a indexed array, a space
  efficient way of indexing into a large array.
}
\keyword{internal}
\alias{indexed_array}
\alias{[[.indexed_array}
\alias{names.indexed_array}
\alias{length.indexed_array}
\arguments{
  \item{env}{environment containing data frame}
  \item{index}{list of indices}
}
