\name{indexed_df}
\title{An indexed list...}
\usage{indexed_df(env, index)}
\description{An indexed list
Create a indexed list, a space efficient way of indexing into a large data frame}
\keyword{internal}
\alias{indexed_df}
\alias{length.indexed}
\alias{names.indexed}
\alias{as.list.indexed}
\alias{[[.indexed_df}
\alias{[.indexed}
\alias{print.indexed}
\arguments{\item{env}{environment containing data frame}
\item{index}{list of indices}}
