% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liply.r
\name{isplit2}
\alias{isplit2}
\title{Split iterator that returns values, not indices.}
\usage{
isplit2(x, f, drop = FALSE, ...)
}
\description{
Split iterator that returns values, not indices.
}
\section{Warning}{
Deprecated, do not use in new code.
}

\seealso{
\code{\link{plyr-deprecated}}
}
\keyword{internal}
