% Generated by roxygen2 (4.0.0): do not edit by hand
\name{join.keys}
\alias{join.keys}
\title{Join keys.
Given two data frames, create a unique key for each row.}
\usage{
join.keys(x, y, by)
}
\arguments{
  \item{x}{data frame}

  \item{y}{data frame}

  \item{by}{character vector of variable names to join by}
}
\description{
Join keys.
Given two data frames, create a unique key for each row.
}
\keyword{internal}

