\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{List to data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Reduce/simplify a list of homogenous objects to a data frame
}
\usage{list_to_dataframe(res, labels = NULLS)}
\arguments{
\item{res}{list of input data}
\item{labels}{a data frame of labels, one row for each element of res}
}

\details{}

\examples{}

