\name{list_to_vector}
\alias{list_to_vector}
\title{List to vector}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Reduce/simplify a list of homogenous objects to a vector
}
\usage{list_to_vector(res)}
\arguments{
\item{res}{list of input data}
}

\details{}

\examples{}
\keyword{internal}
