% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-to-vector.r
\name{list_to_vector}
\alias{list_to_vector}
\title{List to vector.}
\usage{
list_to_vector(res)
}
\arguments{
\item{res}{list of input data}
}
\description{
Reduce/simplify a list of homogenous objects to a vector
}
\seealso{
Other list simplification functions: 
\code{\link{list_to_array}()},
\code{\link{list_to_dataframe}()}
}
\concept{list simplification functions}
\keyword{internal}
