\name{melt.default}
\alias{melt.default}
\title{Melt a vector.}
\usage{\method{melt}{default} (data, ...)}

\description{
  Melt a vector. For vectors, makes a column of a data frame
}
\keyword{manip}
\arguments{
  \item{data}{vector to melt}
  \item{...}{further arguments passed to or from other methods.}
}
