\name{names.quoted}
\alias{names.quoted}
\title{Compute names of quoted variables}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Figure out names of quoted variables, using specified names if they exist, otherwise using \code{\link{make.names}} on the values.
}
\usage{names.quoted(x)}
\arguments{
\item{x}{}
}

\details{}

\examples{}
\keyword{internal}
