% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.r
\name{names.quoted}
\alias{names.quoted}
\title{Compute names of quoted variables.}
\usage{
\method{names}{quoted}(x)
}
\description{
Figure out names of quoted variables, using specified names if they exist,
otherwise converting the values to character strings.  This may create
variable names that can only be accessed using \code{``}.
}
\keyword{internal}
