\name{ninteraction}
\alias{ninteraction}
\title{Numerical interaction}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
A purely numerical interaction function that powers \code{aaply}.
}
\usage{ninteraction(.variables, drop = FALSE)}
\arguments{
\item{.variables}{}
\item{drop}{}
}

\details{}

\examples{}

