\name{ninteraction}
\alias{ninteraction}
\title{Numerical interaction...}
\usage{ninteraction(.variables, drop=FALSE)}

\description{
  Numerical interaction A purely numerical interaction function that
  powers \code{aaply}.
}
\keyword{internal}
\arguments{
  \item{.variables}{list of variables}
  \item{drop}{drop unusued factor levels?}
}
