% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{nunique}
\alias{nunique}
\title{Number of unique values.}
\usage{
nunique(x)
}
\arguments{
\item{x}{vector}
}
\description{
Calculate number of unique values of a variable as efficiently as possible.
}
\keyword{internal}
