\name{print.quoted}
\alias{print.quoted}
\title{Print quoted variables}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Display the \code{\link{str}}ucture of quoted variables
}
\usage{print.quoted(x, ...)}
\arguments{
\item{x}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
