% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.split}
\alias{print.split}
\title{Print split.}
\usage{
\method{print}{split}(x, ...)
}
\arguments{
  \item{x}{object to print}

  \item{...}{unused}
}
\description{
Don't print labels, so it appears like a regular list
}
\keyword{internal}

