% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.r
\name{progress_text}
\alias{progress_text}
\title{Text progress bar.}
\usage{
progress_text(style = 3, ...)
}
\arguments{
\item{style}{style of text bar, see Details section of \code{\link{txtProgressBar}}}

\item{...}{other arugments passed on to \code{\link{txtProgressBar}}}
}
\description{
A textual progress bar
}
\details{
This progress bar displays a textual progress bar that works on all
platforms.  It is a thin wrapper around the built-in
\code{\link{setTxtProgressBar}} and can be customised in the same way.
}
\examples{
l_ply(1:100, identity, .progress = "text")
l_ply(1:100, identity, .progress = progress_text(char = "-"))
}
\seealso{
Other progress bars: 
\code{\link{progress_none}()},
\code{\link{progress_time}()},
\code{\link{progress_tk}()},
\code{\link{progress_win}()}
}
\concept{progress bars}
