\name{rbind.fill}
\alias{rbind.fill}
\title{Rbind fill}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Rbind a list of data frames filling missing columns with NA
}
\usage{rbind.fill(...)}
\arguments{
\item{...}{data frames to row bind together}
}

\details{}

\examples{}
\keyword{manip}
