\name{reorder.character}
\alias{reorder.character}
\title{Reorder character}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Make it possible to reorder character vectors
}
\usage{reorder.character(x, X, FUN = mean, ...)}
\arguments{
\item{x}{}
\item{X}{}
\item{FUN}{}
\item{...}{}
}

\details{Should be removed once this is built into R.}

\examples{}
\keyword{internal}
