\name{split_labels}
\alias{split_labels}
\title{Generate labels for split data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create data frame giving labels for split data frame.
}
\usage{split_labels(splits, drop)}
\arguments{
\item{splits}{list of variables to split up by}
\item{drop}{}
}

\details{}

\examples{}
\keyword{internal}
