\name{split_labels}
\alias{split_labels}
\title{Generate labels for split data frame.}
\usage{split_labels(splits, drop)}

\description{
  Generate labels for split data frame. Create data frame giving labels
  for split data frame.
}
\keyword{internal}
\arguments{
  \item{list}{of variables to split up by}
  \item{whether}{all possible combinations should be considered, or only those present in the data}
}
