% Generated by roxygen2 (4.0.0): do not edit by hand
\name{true}
\alias{true}
\title{Function that always returns true.}
\usage{
true(...)
}
\arguments{
  \item{...}{all input ignored}
}
\value{
\code{TRUE}
}
\description{
Function that always returns true.
}
\seealso{
\code{\link{colwise}} which uses it
}
\keyword{internal}

