% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.r
\name{eval.quoted}
\alias{eval.quoted}
\title{Evaluate a quoted list of variables.}
\usage{
eval.quoted(exprs, envir = NULL, enclos = NULL, try = FALSE)
}
\arguments{
\item{try}{if TRUE, return \code{NULL} if evaluation unsuccesful}

\item{expr}{quoted object to evalution}
}
\value{
a list
}
\description{
Evaluates quoted variables in specified environment
}
\keyword{internal}
