% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexed-array.r
\name{indexed_array}
\alias{indexed_array}
\alias{[[.indexed_array}
\alias{names.indexed_array}
\alias{length.indexed_array}
\title{An indexed array.}
\usage{
indexed_array(env, index)
}
\arguments{
\item{env}{environment containing data frame}

\item{index}{list of indices}
}
\description{
Create a indexed array, a space efficient way of indexing into a large
array.
}
\keyword{internal}
