\name{indexed_df}
\title{An indexed data frame.}
\usage{indexed_df(data, index)}

\description{
  An indexed data frame. Create a indexed list, a space
  efficient way of indexing into a large data frame
}
\keyword{internal}
\alias{indexed_df}
\alias{length.indexed}
\alias{names.indexed}
\alias{as.list.indexed}
\alias{[[.indexed_df}
\alias{[.indexed}
\alias{print.indexed}
\arguments{
  \item{env}{environment containing data frame}
  \item{index}{list of indices}
}
