% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexed-data-frame.r
\name{indexed_df}
\alias{indexed_df}
\title{An indexed data frame.}
\usage{
indexed_df(data, index, vars)
}
\arguments{
\item{data}{environment containing data frame}

\item{index}{list of indices}

\item{vars}{a character vector giving the variables used for subsetting}
}
\description{
Create a indexed list, a space efficient way of indexing into a large data frame
}
\keyword{internal}
