% Generated by roxygen2 (4.0.0): do not edit by hand
\name{indexed_df}
\alias{indexed_df}
\title{An indexed data frame.}
\usage{
indexed_df(data, index, vars)
}
\arguments{
  \item{env}{environment containing data frame}

  \item{index}{list of indices}

  \item{vars}{a character vector giving the variables used
  for subsetting}
}
\description{
Create a indexed list, a space efficient way of indexing into a large data frame
}
\keyword{internal}

