% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.discrete}
\alias{is.discrete}
\title{Determine if a vector is discrete.}
\usage{
is.discrete(x)
}
\arguments{
\item{x}{vector to test}
}
\description{
A discrete vector is a factor or a character vector
}
\examples{
is.discrete(1:10)
is.discrete(c("a", "b", "c"))
is.discrete(factor(c("a", "b", "c")))
}
\keyword{internal}
