\name{join.keys}
\alias{join.keys}
\title{Join keys.}
\usage{join.keys(x, y, by)}

\description{
  Join keys. Given two data frames, create a unique key for each row.
}
\keyword{internal}
\arguments{
  \item{x}{data frame}
  \item{y}{data frame}
  \item{by}{character vector of variable names to join by}
}
