% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.r
\name{join.keys}
\alias{join.keys}
\title{Join keys.
Given two data frames, create a unique key for each row.}
\usage{
join.keys(x, y, by)
}
\arguments{
\item{x}{data frame}

\item{y}{data frame}

\item{by}{character vector of variable names to join by}
}
\description{
Join keys.
Given two data frames, create a unique key for each row.
}
\keyword{internal}
