% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liply.r
\name{liply}
\alias{liply}
\title{Experimental iterator based version of llply.}
\usage{
liply(.iterator, .fun = NULL, ...)
}
\arguments{
\item{.iterator}{iterator object}

\item{.fun}{function to apply to each piece}

\item{...}{other arguments passed on to \code{.fun}}
}
\description{
Because iterators do not have known length, \code{liply} starts by
allocating an output list of length 50, and then doubles that length
whenever it runs out of space.  This gives O(n ln n) performance rather
than the O(n ^ 2) performance from the naive strategy of growing the list
each time.
}
\section{Warning}{
Deprecated, do not use in new code.
}

\seealso{
\code{\link{plyr-deprecated}}
}
\keyword{manip}
