% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-to-dataframe.r
\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{List to data frame.}
\usage{
list_to_dataframe(res, labels = NULL, id_name = NULL, id_as_factor = FALSE)
}
\arguments{
\item{res}{list of input data}

\item{labels}{a data frame of labels, one row for each element of res}

\item{idname}{the name of the index column, \code{NULL} for no index
column}
}
\description{
Reduce/simplify a list of homogenous objects to a data frame. All
\code{NULL} entries are removed. Remaining entries must be all atomic
or all data frames.
}
\seealso{
Other list simplification functions: 
\code{\link{list_to_array}()},
\code{\link{list_to_vector}()}
}
\concept{list simplification functions}
\keyword{internal}
