% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plyr-deprecated.r
\name{plyr-deprecated}
\alias{plyr-deprecated}
\title{Deprecated Functions in Package plyr}
\description{
These functions are provided for compatibility with older versions of
\code{plyr} only, and may be defunct as soon as the next release.
}
\details{
\itemize{
  \item \code{\link{liply}}
  \item \code{\link{isplit2}}
}
}

