% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{print.split}
\alias{print.split}
\title{Print split.}
\usage{
\method{print}{split}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{unused}
}
\description{
Don't print labels, so it appears like a regular list
}
\keyword{internal}
