% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.r
\name{progress_win}
\alias{progress_win}
\title{Graphical progress bar, powered by Windows.}
\usage{
progress_win(title = "plyr progress", ...)
}
\arguments{
\item{title}{window title}

\item{...}{other arguments passed on to \code{winProgressBar}}
}
\description{
A graphical progress bar displayed in a separate window
}
\details{
This graphical progress only works on Windows.
}
\examples{
if(exists("winProgressBar")) {
l_ply(1:100, identity, .progress = "win")
l_ply(1:100, identity, .progress = progress_win(title="Working..."))
}
}
\seealso{
\code{winProgressBar} for the function that powers this progress bar

Other progress bars: 
\code{\link{progress_none}()},
\code{\link{progress_text}()},
\code{\link{progress_time}()},
\code{\link{progress_tk}()}
}
\concept{progress bars}
