% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_indices}
\alias{split_indices}
\title{Split indices.}
\usage{
split_indices(group, n = 0L)
}
\arguments{
\item{n}{largest integer (may not appear in index). This is hint: if
the largest value of \code{group} is bigger than \code{n}, the output
will silently expand.}

\item{index}{integer indices}
}
\description{
An optimised version of split for the special case of splitting row
indices into groups, as used by \code{\link{splitter_d}}.
}
\examples{
split_indices(sample(10, 100, rep = TRUE))
split_indices(sample(10, 100, rep = TRUE), 10)
}
\keyword{internal}
\keyword{manip}
