% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter-d.r
\name{split_labels}
\alias{split_labels}
\title{Generate labels for split data frame.}
\usage{
split_labels(splits, drop, id = plyr::id(splits, drop = TRUE))
}
\arguments{
\item{list}{of variables to split up by}

\item{whether}{all possible combinations should be considered, or only those present in the data}
}
\description{
Create data frame giving labels for split data frame.
}
\keyword{internal}
