% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.r
\name{try_default}
\alias{try_default}
\alias{tryNULL}
\title{Try, with default in case of error.}
\usage{
try_default(expr, default, quiet = FALSE)

tryNULL(expr)
}
\arguments{
\item{expr}{expression to try}

\item{default}{default value in case of error}

\item{quiet}{should errors be printed (TRUE) or ignored (FALSE, default)}
}
\description{
\code{try_default} wraps try so that it returns a default value in the case of error.
\code{tryNULL} provides a useful special case when dealing with lists.
}
\seealso{
\code{\link{tryapply}}
}
\keyword{internal}
